pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llavemx-mobile-bridge"
version = "0.1.0"
description = "Plugin de Open edX para autenticación móvil con LlaveMX (compatible con Ulmo)"
readme = "README.md"
license = {text = "AGPL-3.0"}
requires-python = ">=3.8"
authors = [
    {name = "Tu Nombre", email = "tu@email.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Django",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "Django>=3.2",
    "djangorestframework",
    "requests",
    "edx-django-utils",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-django",
]

[project.urls]
Homepage = "https://github.com/tu-usuario/llavemx_mobile_bridge"
Repository = "https://github.com/tu-usuario/llavemx_mobile_bridge"

[project.entry-points."lms.djangoapp"]
llavemx_mobile_bridge = "llavemx_mobile_bridge.apps:LlaveMxMobileBridgeConfig"

[tool.setuptools.packages.find]
where = ["."]
include = ["llavemx_mobile_bridge*"]
